+++ nodes/__init__.py
@@ -1,6 +1,6 @@
 
 """ Node classes used by the Kasa Node Server. """
-VERSION = "3.3.4"
+VERSION = "3.3.5"
 
 import udi_interface
 #import sys